+++ rpcs3/Emu/CMakeLists.txt
@@ -437,6 +437,7 @@ if(TARGET 3rdparty_vulkan)
 	target_sources(rpcs3_emu PRIVATE
 		RSX/VK/vkutils/buffer_view.cpp
 		RSX/VK/vkutils/chip_class.cpp
+		RSX/VK/vkutils/command_pool.cpp
 		RSX/VK/vkutils/fence.cpp
 		RSX/VK/vkutils/mem_allocator.cpp
 		RSX/VK/vkutils/memory_block.cpp
