+++ rpcs3/VKGSRender.vcxproj
@@ -42,6 +42,7 @@
     <ClInclude Include="Emu\RSX\VK\VKTextureCache.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\buffer_view.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\chip_class.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\command_pool.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\fence.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\graphics_pipeline_state.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\memory_block.h" />
@@ -77,6 +78,7 @@
     <ClCompile Include="Emu\RSX\VK\VKTexture.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\buffer_view.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\chip_class.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\command_pool.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\fence.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\memory_block.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\mem_allocator.cpp" />
