+++ rpcs3/VKGSRender.vcxproj.filters
@@ -50,6 +50,9 @@
     <ClCompile Include="Emu\RSX\VK\vkutils\render_device.cpp">
       <Filter>vkutils</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\RSX\VK\vkutils\command_pool.cpp">
+      <Filter>vkutils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Emu\RSX\VK\VKCommonDecompiler.h" />
@@ -112,6 +115,9 @@
       <Filter>vkutils</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\VK\vkutils\render_device.h">
+      <Filter>vkutils</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\VK\vkutils\command_pool.h">
       <Filter>vkutils</Filter>
     </ClInclude>
   </ItemGroup>
